requirements.txt
python-binance==0.7.5
pandas
numpy
APScheduler==3.2.0
